Adding user accounts.

+++ PHPCI/Controller/BuildController.php
@@ -59,6 +59,11 @@ class BuildController extends b8\Controller
 
 	public function delete($buildId)
 	{
+		if(!Registry::getInstance()->get('user')->getIsAdmin())
+		{
+			throw new \Exception('You do not have permission to do that.');
+		}
+		
 		$build	= $this->_buildStore->getById($buildId);
 		$this->_buildStore->delete($build);
